create note functional and auto renders
